+++ pyskoob/__init__.py
@@ -9,7 +9,7 @@ from .auth import AuthService
 from .authors import AuthorService
 from .books import BookService
 from .client import SkoobClient
-from .exceptions import ParsingError
+from .exceptions import HTTPClientError, ParsingError
 from .http.httpx import HttpxAsyncClient, HttpxSyncClient
 from .profile import SkoobProfileService
 from .publishers import PublisherService
@@ -25,5 +25,6 @@ __all__ = [
     "SkoobClient",
     "SkoobProfileService",
     "UserService",
+    "HTTPClientError",
     "ParsingError"
 ]